refactor main file, to have a file with a class of options
